backend/requirements.txt
fastapi
uvicorn
tensorflow
pillow
numpy
pymongo
cloudinary
python-dotenv
scikit-learn
python-multipart
requests
PyJWT
bcrypt
email-validator


